+++ src/BioSequences.jl
@@ -140,9 +140,6 @@ export
     RNACodon,
     translate,
     ncbi_trans_table,
-    FASTA,
-    FASTQ,
-    TwoBit,
     ABIF,
     
     # Indexing
@@ -233,12 +230,6 @@ include("iterators/eachkmer.jl")
 include("composition.jl")
 include("geneticcode.jl")
 include("demultiplexer.jl")
-
-# Parsing file types
-include("fasta/fasta.jl")
-include("fastq/fastq.jl")
-include("twobit/twobit.jl")
-include("abif/abif.jl")
 
 include("search/exact.jl")
 include("search/approx.jl")