+++ test/runtests.jl
@@ -213,11 +213,4 @@ include("translation.jl")
 
 include("demultiplexer.jl")
 
-@testset "Reading and Writing" begin
-    include("io/FASTA.jl")
-    include("io/FASTQ.jl")
-    include("io/twobit.jl")
-    include("io/abif.jl")
-end
-
 end